+++ KoS.Apps.SharePoint.SmartCAML/changelog.md
@@ -4,7 +4,8 @@
 ##### Update
 + Change the description for API section in Connect window
 + Read order by settings from xml to designer
-+ new options for items grid header context menu: "Hide all hidden" and "Hide all readonly"
++ New options for items grid header context menu: "Hide all hidden" and "Hide all readonly"
++ New options for items grid header context menu: "Pin" and "Unpin"
 
 ##### Fix
 + Change the watermark for password input in Connect window